requirements.txt
streamlit
langchain
langchain-community
fastapi
uvicorn
python-dotenv
requests
plotly
faiss-cpu
openai
langchain-openai
tabulate
numpy